Move pages into folders and add css
