+++ frontend/src/pages/HomePage/HomePage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import DixieTechLogo from "../assets/DixieTechLogo.png";
+import DixieTechLogo from "../../assets/DixieTechLogo.png";
 
 const Home: React.FC = () => {
   return (
