+++ frontend/src/pages/RxItem/RxItem.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import SearchBar from "../components/SearchBar";
+import SearchBar from "../../components/SearchBar";
 
 const RxItem: React.FC = () => {
   const navigate = useNavigate();
